refactor: 입력 문자열에서 Integer 리스트로 변환하는 과정 클래스 추출
